add settings page & split sidebar component
